cli-monitor-v2.1/requirements.txt
python-dotenv
python-binance
rich
